+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 <br/>
 
 <p>
+<b>Version 1.1.98</b> - December 2020<br/>
+- NEW: added Datapoint 242.600 Color xyY.<br/>
+</p>
+<p>
 <b>Version 1.1.97</b> - December 2020<br/>
 - NEW: added help links directly into the config windows and upon selection of datapoint as well.<br/>
 - Cleaning of UI.<br/>